restaurantManagement/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.1
cloudinary==1.41.0
cryptography==43.0.1
Django==5.1
django-ckeditor==6.7.1
django-form-surveys==2.4.0
django-js-asset==2.2.0
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
pytz==2024.1
PyYAML==6.0.2
rest-framework-simplejwt==0.0.2
setuptools==74.1.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
